// Models/SuiviAgenda.model.ts
export interface SuiviAgenda {
  num?: number;
  date_ag?: Date;
  objet?: string;
  app_gen?: string;
  sort?: string;
  compte_rendu?: string;
  mnt_reg?: number;
  date_ver?: Date;
  lieu_ver?: string;
  mnt_plan?: number;
  nb_ech?: number;
  date_prem_ver?: Date;
  date_rdv?: Date;
  type_rdv?: string;
  date_visite?: Date;
  lieu_visite?: string;
  eng?: string;
  observation?: string;
  aff?: string;
  ges?: string;
  mnt_rec?: number;
  dat_rec?: Date;
  lieu_rec?: string;
  nouv_tel?: string;
  nouv_te2?: string;
  nouv_fax?: string;
  nouv_adresse?: string;
  nature?: string;
  h_rdv?: string;
  matricule_rdv?: string;
  ref?: number;
  date_deuxi_ech?: Date;
  date_trois_ech?: Date;
  mntech1?: number;
  mntech2?: number;
  mntech3?: number;
  liste_choix?: string;
  lieu?: string;
  date_l?: Date;
  taux?: string;
  mnt_rec2?: number;
  mnt_rec3?: number;
  code_fonction?: number;
  type_lettre?: string;
  date_lettre?: Date;
  cli?: string;
  mntech4?: number;
  mntech5?: number;
  date_quat_ech?: Date;
  date_cinq_ech?: Date;
  motif_imp?: string;
  info_motif?: string;
  date_motif?: Date;
  versement_prealable?: number;
  agc?: string;
  date_action?: Date;
  usr_nom?: string;
  usr_matricule?: string;
  id: number;
  montantFacilites?: any[];
}