// Models/AbCompte.model.ts
import { ab_client } from "./ab_client.model";

export interface AbCompte {
  cli?: string;
  age?: string;
  ncp?: string;
  classe_imp?: number;
  classe_gel?: number;
  montant_aut?: number;
  mnt_dep_aut?: number;
  nbre_imp?: number;
  mnt_imp?: number;
  nbre_defaut?: number;
  nom?: string;
  nbre_d_aut?: number;
  date_auto?: Date;
  date_finale?: Date;
  nombre_jours?: number;
  marche?: string;
  groupe?: string;
  etat_lettre?: string;
  date_lettre_recouv?: Date;
  etat_carte?: string;
  date_carte?: Date;
  etat_lmd?: string;
  date_lmd?: Date;
  matricule_lettre?: string;
  lettre_employeur?: string;
  date_lettre_employeur?: Date;
  etat_susp?: string;
  date_susp?: Date;
  mnt_sdb?: number;
  matricule_ls?: string;
  matricule_lmd?: string;
  flag_ls?: string;
  flag_lmd?: string;
  lmd1_prop?: string;
  date_lmd2?: Date;
  matricule_lmd2?: string;
  etat_lmd2?: string;
  lmd2_prop?: string;
  date_prop_lmd1?: Date;
  matricule_prop_lmd1?: string;
  date_prop_lmd2?: Date;
  matricule_prop_lmd2?: string;
  susp_lmd1?: string;
  susp_lmd2?: string;
  susp_lr?: string;
  tot_eng?: number;
  phase?: string;
  region?: string;
  flag_lmd2?: string;
  nombre_jours_sdb?: number;
  depassement?: number;
  tot_creance?: number;
  max_nbj?: number;
  flg_gen_ltrd?: string;
  flg_ltrd?: string;
  flg_ltrd_date?: Date;
  prop_trf_rd?: string;
  prop_trf_rd_date?: Date;
  prop_trf_rd_matricule?: string;
  val_trf_rd?: string;
  val_trf_rd_date?: Date;
  val_trf_rd_matricule?: string;
  code_rec?: string;
  AB_COMPTE_ID: number;
  ab_client: ab_client;
}